set Mt size half of the figure
