+++ .gitignore
@@ -13,3 +13,5 @@ prep_files/biomaRt.R
 sequences.fasta
 temp_files/regions.csv
 Fig1.png
+Figure1_after_rev.png
+MANUDBprocess.png
